feat: refactor app structure to use routers and initialize services in common.py
